third_party/blink/web_tests/fast/history/same-document-iframes-changing-pushstate-expected.txt

Tests that we trigger same-document navigation when history entries are generated via pushState, even if the frames present in the document change between history entries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

removing iframe2
pushState to newState
going back
popstate to state: null filename: same-document-iframes-changing-pushstate.html
going forward
popstate to state: newState filename: same-document-iframes-changing-pushstate.html?newState
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'iframe1'
--------
iframe1

============== Back Forward List ==============
        (file test):fast/history/same-document-iframes-changing-pushstate.html
            data:text/plain,iframe1 (in frame "iframe1")
            data:text/plain,iframe2 (in frame "iframe2")
curr->  (file test):fast/history/same-document-iframes-changing-pushstate.html?newState
            data:text/plain,iframe1 (in frame "iframe1")
===============================================
